requirements.txt
pandas
numpy
alpaca-py
requests
lxml
pytz
joblib
yfinance
pyarrow
fastparquet
scipy
